pyproject.toml
[project]
name = "4o-ghibli-at-home"
version = "0.1.9"
description = "GPT-4o Ghibli At Home – Private AI photo stylizer powered by FLUX Kontext"
requires-python = ">=3.11"
authors = [
    { name = "Ahmad M. Osman", email = "me@ahmadosman.com" }
]
license = { file = "LICENSE" }
readme = "README.md"
keywords = [
    "image-generation", "ai", "photo-stylizer", "ghibli", "diffusers", "vram-optimized", "ml", "art", "flux-kontext", "dfloat11", "private-ai", "local-llm", "local-ai"
]
classifiers = [
    "Intended Audience :: End Users/Desktop",
    "Intended Audience :: End Users/Web-Based",
    "License :: OSI Approved :: GNU Affero General Public License v3 or later (AGPLv3+)",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Operating System :: POSIX :: Linux",
    "Operating System :: Microsoft :: Windows",
    "Topic :: Multimedia :: Graphics :: Graphics Conversion",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Development Status :: 3 - Alpha"
]

dependencies = [
    "celery",
    "redis",
    "dfloat11[cuda12]>=0.2.0",
    "diffusers",
    "flask>=3.1.1",
    "flask-cors>=6.0.1",
    "loguru>=0.7.3",
    "protobuf>=6.31.1",
    "python-dotenv>=1.1.1",
    "sentencepiece>=0.2.0",
]

[dependency-groups]
prod = [
    "gunicorn>=23.0.0",
]

[project.optional-dependencies]
dev = [
    "pre-commit>=4.2.0",
    "ruff>=0.12.2",
    "mypy>=1.16.1",
]

[tool.uv]
default-groups = []

[tool.uv.sources]
diffusers = { git = "https://github.com/huggingface/diffusers" }

[tool.ruff]
extend-include = ["*.ipynb"]
line-length = 90
target-version="py312"

[tool.ruff.lint]
select = ["E", "F", "UP", "B", "SIM", "I", "W", "Q"]
ignore = ["E203", "E501"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
docstring-code-format = true
docstring-code-line-length = 80
